Add pos-only args, more type narrowing

+++ neofoodclub/utils.py
@@ -80,7 +80,7 @@ class Table:
         return "\n".join(lines)
 
 
-def fix_bet_amounts(amts: npt.NDArray[np.int_]) -> npt.NDArray[np.int_]:
+def fix_bet_amounts(amts: npt.NDArray[np.int32]) -> npt.NDArray[np.int32]:
     """:class:`np.ndarray`: Returns a "clamped" array of the bet amounts passed in where the minimum value is 50 and
     the maximum value is 70304, which is the highest value that the current hashing algorithm can understand."""
     # fix any values below 50 to be 50, to maintain working bets
